foxsense-satellite/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pystac-client==0.8.3
planetary-computer==1.0.0
rasterio==1.3.10
numpy==1.26.4
scipy==1.13.1
scikit-image==0.24.0
Pillow==10.4.0
matplotlib==3.9.1
shapely==2.0.5
pyproj==3.6.1
httpx==0.27.0
python-dotenv==1.0.1
